refactor: extract HTTP handler and backend contracts from server.ts (AC0-AC4)
Batch B server decomposition — extract cohesive modules from the monolithic
createBackendServer() closure:

- AC0: Add CurateHandle.stop() and clearAll(registry) prerequisites
- AC1: Extract routeToolCall() into execution/tool-router.ts
- AC2: Extract KB tool handling into execution/kb-tools.ts
- AC3: Extract discuss tool handling into execution/discuss-tools.ts
- AC4: Extract HTTP request handling into execution/http-handler.ts with
  explicit dependency contracts in execution/backend-contracts.ts

AC4 defines BackendIdentity, ReadonlyBackendRuntimeState,
MutableBackendRuntimeState, HttpHandlerDeps, and EventStreamHandlers
contracts so the HTTP handler receives all dependencies through injection
rather than hidden closure reads.

Behavior fix: drain admission fence now rejects new work-admitting requests
immediately after /admin/shutdown, before lifecycle transitions to 'draining',
closing a pre-existing race window.

server.ts drops from 1591 to 1149 lines; remaining lifecycle/recovery code
will move in AC5.
